+++ core/vector3d.cpp
@@ -1,8 +1,8 @@
-#include <iostream>
+#include <stdexcept>
 #include <cmath>
 #include "vector3d.h"
     
-    vector3d::vector3d(float x_val, float y_val, float z_val) {
+    vector3d::vector3d(double x_val, double y_val, double z_val) {
         x = x_val;
         y = y_val;
         z = z_val;  
@@ -18,19 +18,19 @@ vector3d vector3d::sub(vector3d other) const {
     return result;
 }
 
-vector3d vector3d::mult(float scalar) const {
+vector3d vector3d::mult(double scalar) const {
     vector3d result = vector3d(x * scalar, y * scalar, z * scalar);
     return result;
 }
 
-float vector3d::magnitude() const {
-    float result = std::sqrt(x*x + y*y + z*z);
+double vector3d::magnitude() const {
+    double result = std::sqrt(x*x + y*y + z*z);
     return result;
     
 }
 
 vector3d vector3d::normalized() const {
-    float mag = magnitude();
+    double mag = magnitude();
     if (mag == 0) {
         throw std::invalid_argument("Can't normalize a zero vector.");
     }
@@ -38,8 +38,8 @@ vector3d vector3d::normalized() const {
     return result;
 }
 
-float vector3d::dot(vector3d other) const {
-    float result = x*other.x + y*other.y + z*other.z;
+double vector3d::dot(vector3d other) const {
+    double result = x*other.x + y*other.y + z*other.z;
     return result;
 }
 
@@ -51,6 +51,6 @@ vector3d vector3d::cross(const vector3d& other) const {
     );
 }
 
-vector3d lerp(vector3d a, vector3d b, float t) {
+vector3d lerp(vector3d a, vector3d b, double t) {
     return a.add(b.sub(a).mult(t));
 }
